requirements.txt
scipy
numpy
resampy
panphon
torch
editdistance
requests
tqdm
allosaurus
elevenlabs
Flask
tabulate
pydub
pyfestival
